common/syscalls: Moved all the system calls implementations into their own file
