+++ src/platforms/common/Makefile.inc
@@ -27,4 +27,5 @@ SRC +=             \
 	usb.c          \
 	usb_serial.c   \
 	usb_dfu_stub.c \
-	aux_serial.c
+	aux_serial.c   \
+	syscalls.c
